Moved the list transaction api called to the api mod
