+++ src/transaction/mod.rs
@@ -13,17 +13,8 @@ mod api;
 
 pub use self::api::get_transaction_v2;
 pub use self::api::delete_transaction;
+pub use self::api::list_transactions;
 
-pub async fn list_transactions(
-    pool: web::Data<Pool<SqliteConnectionManager>>,
-) -> Result<HttpResponse, Error> {
-    let result = db::list_transactions(pool.get().unwrap());
-
-    match result {
-        Ok(v) => Ok(HttpResponse::Ok().json(v)),
-        Err(_e) => Ok(HttpResponse::InternalServerError().finish()),
-    }
-}
 
 pub async fn list_transactions_with_details(
     pool: web::Data<Pool<SqliteConnectionManager>>,
